requirement.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
pycparser==2.19
PyScons==1.0.95
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
python-telegram-bot==12.3.0
requests==2.7.0
six==1.14.0
telebot==0.0.3
tornado==6.0.3
urllib3==1.25.8
